requirements.txt
blinker==1.9.0
certifi==2026.1.4
click==8.3.1
cloudinary==1.44.1
Flask==3.1.2
Flask-Login==0.6.3
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
six==1.17.0
urllib3==2.6.3
Werkzeug==3.1.5
